**Title:** v0.5.13 Template Merged

**Description:**
Fixes the navigation issues when routing to Logout, Settings, or clicking UserAvatar.
Dallas is he real MVP.

**v0.5.13 Changes:**
- Fixes the issue with routing to Home when click UserAvatar
- Adds hasException checks to other Nav helpers
- Fixes routing behavior during logout
- Calling init during login now
- - Cleaned up the onInit hook
- Cleans up waitForStorageToBeReady call and mult auth calls
- Added lifecycle diagram
- Persists darkMode even after logging out
- Inheritting other versions changes before v0.5.13

+++ scripts/stop_test_servers.sh
@@ -26,7 +26,7 @@ for containerName in "${ALL_ContainerArray[@]}"
 do
   if docker container ls -a | grep -Fq "$containerName" 1>/dev/null; then
      # If container name starts with "mt-", then it's an MT object and must be removed
-     if [[ $containerName = "mt-"* ]]; then
+     if [[ $containerName == "mt-"* ]] && [[ $containerName != "mt-ingress" ]]; then
         echo docker container rm --force ${containerName}
         docker container rm --force ${containerName}
      else
